Add check-ready action

+++ 5.10/docker-entrypoint.sh
@@ -3,23 +3,25 @@
 set -e
 
 if [[ -n "${DEBUG}" ]]; then
-  set -x
+    set -x
 fi
 
-chown daemon:daemon /var/atlassian/confluence
+if [[ "${1}" == 'make' ]]; then
+    su-exec daemon "${@}" -f /usr/local/bin/actions.mk
+else
+    chown daemon:daemon /var/atlassian/confluence
 
-if [[ -n "${JIRA_HOST}" ]]; then
-    jre="/usr/lib/jvm/java-1.8-openjdk/jre"
+    if [[ -n "${JIRA_HOST}" ]]; then
+        if [[ -z "${JIRA_PORT}" ]]; then
+            JIRA_PORT=443
+        fi
 
-    if [[ -z "${JIRA_PORT}" ]]; then
-        JIRA_PORT=443
+        url="${JIRA_HOST}:${JIRA_PORT}"
+        openssl s_client -connect "${url}"  < /dev/null | sed -ne '/-BEGIN CERTIFICATE-/,/-END CERTIFICATE-/p' > /tmp/cert.pem
+        "${JAVA_HOME}/bin/keytool" -import -alias jira -trustcacerts -keystore "${JAVA_HOME}/lib/security/cacerts" \
+            -file /tmp/cert.pem -storepass changeit -noprompt
+        rm /tmp/cert.pem
     fi
 
-    url="${JIRA_HOST}:${JIRA_PORT}"
-    openssl s_client -connect "${url}"  < /dev/null | sed -ne '/-BEGIN CERTIFICATE-/,/-END CERTIFICATE-/p' > /tmp/cert.pem
-    "${jre}/bin/keytool" -import -alias jira -trustcacerts -keystore "${jre}/lib/security/cacerts" \
-        -file /tmp/cert.pem -storepass changeit -noprompt
-    rm /tmp/cert.pem
+    su-exec daemon "${@}"
 fi
-
-su-exec daemon "${@}"
